server/pyproject.toml
[tool.poetry]
name = "chatbot-widget-server"
version = "0.1.0"
description = "A chatbot widget backend using FastAPI"
authors = ["rajharsh81070 <rajharsh81070@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
sqlalchemy = "^2.0.31"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
